# shared/api/openapi.yaml
openapi: 3.0.3
info:
  title: Ada Fitness Tracker API
  description: API for the Ada fitness and nutrition tracking application
  version: 1.0.0
  contact:
    email: contact@ada-fitness.app
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.ada-fitness.app
    description: Production server
...
paths:
  /auth/magic-link:
    post:
      tags: [Authentication]
      summary: Request magic link
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
              required: [email]
      responses:
        '200':
          description: Magic link sent
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  debug_link:
                    type: string
  /auth/verify:
    post:
      tags: [Authentication]
      summary: Verify magic link token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                token:
                  type: string
              required: [token]
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
  /auth/refresh:
    post:
      tags: [Authentication]
      summary: Refresh access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                refresh_token:
                  type: string
              required: [refresh_token]
      responses:
        '200':
          description: Token refreshed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
...
  /ai/chat:
    post:
      tags: [AI]
      summary: Chat with Ada assistant
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                message:
                  type: string
                conversation_id:
                  type: string
                include_context:
                  type: boolean
                  default: true
              required: [message]
      responses:
        '200':
          description: Chat response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatResponse'
  /ai/vision/analyze:
    post:
      tags: [AI]
      summary: Analyze food photo
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                image_base64:
                  type: string
                prompt:
                  type: string
              required: [image_base64]
      responses:
        '200':
          description: Vision analysis result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VisionAnalyzeResponse'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    TokenResponse:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        token_type:
          type: string
        expires_in:
          type: integer
        user_id:
          type: string
...
    ChatResponse:
      type: object
      properties:
        message:
          type: string
        role:
          type: string
        tool_calls:
          type: array
          items:
            type: object
        conversation_id:
          type: string
        model_used:
          type: string
        created_entries:
          type: array
          items:
            type: object
    VisionAnalyzeResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/VisionFoodItem'
        totals:
          type: object
          properties:
            calories:
              type: integer
            protein_g:
              type: number
            carbs_g:
              type: number
            fat_g:
              type: number
        confidence:
          type: number
        description:
          type: string
        disclaimer:
          type: string
    VisionFoodItem:
      type: object
      properties:
        name:
          type: string
        portion_description:
          type: string
        grams_estimate:
          type: number
        calories:
          type: integer
        protein_g:
          type: number
        carbs_g:
          type: number
        fat_g:
          type: number
        confidence:
          type: number